sensor-exporter/requirements.txt
prometheus_client
psycopg2-binary
python-dotenv
paho-mqtt
loguru
pathlib